tests(cli): move sim soak parse tests out of args module
